Refactor Package for Improved Discoverability and Maintainability #20
